+++ config/content/Footer.js
